# docs/swagger.yaml
basePath: /api/v1
definitions:
  core_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short-human readable message
        type: string
    type: object
...
  tasks_transport_http.CreateTaskRequest:
    properties:
      description:
        description: Описание задачи. Опционально; если передано — от 3 до 1000 символов.
        example: 2 литра, до пятницы
        maxLength: 1000
        minLength: 3
        type: string
      title:
        description: Заголовок задачи. Обязателен, от 1 до 100 символов.
        example: Купить молоко
        maxLength: 100
        minLength: 1
        type: string
      user_id:
        description: ID пользователя-автора задачи. Обязателен.
        example: 1
        type: integer
    required:
    - title
    - user_id
    type: object
  tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-06-30T18:30:00Z"
        type: string
      created_at:
        example: "2026-06-30T12:00:00Z"
        type: string
      description:
        example: 2 литра, до пятницы
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 1
        type: integer
    type: object
  tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-06-30T18:30:00Z"
        type: string
      created_at:
        example: "2026-06-30T12:00:00Z"
        type: string
      description:
        example: 2 литра, до пятницы
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 1
        type: integer
    type: object
  tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        description: Признак завершённости. Поле опционально; null недопустим.
        example: true
        type: boolean
      description:
        description: Новое описание. Поле опционально; передача null очищает описание
          в БД.
        example: 2 литра, до пятницы
        type: string
        x-nullable: true
      title:
        description: Новый заголовок. Поле опционально; null недопустим. От 1 до 100
          символов.
        example: Купить молоко
        type: string
    type: object
  tasks_transport_http.PatchTaskResponse:
    properties:
      author_user_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-06-30T18:30:00Z"
        type: string
      created_at:
        example: "2026-06-30T12:00:00Z"
        type: string
      description:
        example: 2 литра, до пятницы
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 1
        type: integer
    type: object
  tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-06-30T18:30:00Z"
        type: string
      created_at:
        example: "2026-06-30T12:00:00Z"
        type: string
      description:
        example: 2 литра, до пятницы
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 1
        type: integer
    type: object
  users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivanov Ivan
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+79998887766"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivanov Ivan
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 1
        type: integer
    type: object
...
  users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivanov Ivan
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 1
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: kek
  title: Go rest prod API
  version: "1.0"
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: Получить список задач с пагинацией и опциональной фильтрацией по
        автору
      parameters:
      - description: Лимит записей
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      - description: Фильтр по ID автора задачи
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список задач
          schema:
            items:
              $ref: '#/definitions/tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Получить список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу в системе
      parameters:
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданная задача
          schema:
            $ref: '#/definitions/tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удалить задачу из системы по ID
      parameters:
      - description: ID удаляемой задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешное удаление
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Удалить задачу
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Получить задачу по ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Найденная задача
          schema:
            $ref: '#/definitions/tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Получить задачу
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующей в системе задаче.
        ### Логика обновления полей (three-state logic):
        1. **Поле не передано**: значение в БД не меняется
        2. **Явно передано значение**: `"description": "новый текст"` — устанавливает новое значение
        3. **Передан null**: `"description": null` — очищает поле в БД (set to NULL)
        Ограничения: `title` и `completed` не могут быть выставлены как null
      parameters:
      - description: ID изменяемой задачи
        in: path
        name: id
        required: true
        type: integer
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённая задача
          schema:
            $ref: '#/definitions/tasks_transport_http.PatchTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Изменение задачи
      tags:
      - tasks
  /users:
    get:
      consumes:
      - application/json
      description: Получить список пользователей с пагинацией
      parameters:
      - description: Лимит записей
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Получить список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
...
swagger: "2.0"